Migrated /server to the root directory, merged server/package.json with the root package.json, updated root .gitignore to Github's Node.gitignore, updated the server to use .env rather than nodemon.json, and updated server to be ran with ts-node in order to remove the need to have a /dist compile directory for the server code
