+++ package.json
@@ -4,10 +4,13 @@
   "description": "",
   "main": "server/dist/server.js",
   "scripts": {
-    "start": "node server/dist/server.js",
-    "develop": "concurrently \"cd server && npm run watch\" \"cd client && npm start\"",
-    "install": "cd server && npm i && cd ../client && npm i",
-    "build": "cd server && npm run build && cd ../client && npm run build"
+    "start": "ts-node -r dotenv/config server.ts",
+    "watch-local": "concurrently \"nodemon server.ts\" \"cd client && npm start\"",
+    "watch-dev": "concurrently \"nodemon -r dotenv/config server.ts\" \"cd client && npm start\"",
+    "install": "npm i && cd client && npm i",
+    "build": "cd client && npm run build",
+    "seed-local": "ts-node seeders/userSeed.ts",
+    "seed-dev": "ts-node -r dotenv/config seeders/userSeed.ts"
   },
   "repository": {
     "type": "git",
@@ -21,6 +24,16 @@
   },
   "homepage": "https://github.com/yujuncho7/IdoctorReact#readme",
   "devDependencies": {
-    "concurrently": "^6.2.0"
+    "concurrently": "^6.2.0",
+    "dotenv": "^10.0.0",
+    "nodemon": "^2.0.12",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5"
+  },
+  "dependencies": {
+    "@types/express": "^4.17.13",
+    "@types/node": "^16.4.7",
+    "express": "^4.17.1",
+    "mongoose": "^5.13.5"
   }
 }
