--- server/.gitignore
@@ -1,29 +1,0 @@
-# OS X
-.DS_Store*
-Icon?
-._*
-
-# Windows
-Thumbs.db
-ehthumbs.db
-Desktop.ini
-
-# Linux
-.directory
-*~
-
-
-# npm
-node_modules
-package-lock.json
-*.log
-*.gz
-
-# production
-/dist
-
-# Coveralls
-coverage
-
-# Benchmarking
-benchmarks/graphs
