--- server/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "start": "node dist/server.js",
-    "watch": "concurrently \"npx tsc -w\" \"nodemon dist/server.js\"",
-    "build": "npx tsc",
-    "seed-local": "npx tsc && node dist/seeders/userSeed.js",
-    "seed-dev": "npx tsc && nodemon dist/seeders/userSeed.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@types/bcryptjs": "^2.4.2",
-    "@types/express": "^4.17.13",
-    "@types/mongoose": "^5.11.97",
-    "@types/node": "^16.4.0",
-    "@types/passport": "^1.0.7",
-    "bcryptjs": "^2.4.3",
-    "express": "^4.17.1",
-    "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.13.3",
-    "passport": "^0.4.1",
-    "passport-jwt": "^4.0.0",
-    "passport-local": "^1.0.0",
-    "typescript": "^4.3.5"
-  },
-  "devDependencies": {
-    "concurrently": "^6.2.0",
-    "nodemon": "^2.0.12"
-  }
-}
